+++ test/applyiter.jl
@@ -502,4 +502,3 @@ end
     @test ddata[1][1] ≈ 6.0
 end
 
-include("mixedapplyiter.jl")
