+++ test/ext/bfloat16s.jl
@@ -2,6 +2,9 @@ using Enzyme
 using Test
 using BFloat16s
 
-@test_broken Enzyme.gradient(Reverse, sum, ones(BFloat16, 10)) ≈ ones(BFloat16, 10)
+using Enzyme: gradient
 
-@test_broken Enzyme.gradient(Forward, sum, ones(BFloat16, 10)) ≈ ones(BFloat16, 10)
+@testset "BFloat16s ext" begin
+    @test_broken gradient(Reverse, sum, ones(BFloat16, 10)) ≈ ones(BFloat16, 10)
+    @test_broken gradient(Forward, sum, ones(BFloat16, 10)) ≈ ones(BFloat16, 10)
+end
