+++ test/utils.jl
@@ -1,0 +1,68 @@
+#====================================================================================================
+    utils.jl
+====================================================================================================#
+using Enzyme
+
+# symbol is \simeq
+# this is basically a more flexible version of ≈
+(≃)(a, b) = (≈)(a, b)
+(≃)(a::Tuple, b::Tuple) = all(xy -> xy[1] ≃ xy[2], zip(a,b))
+function (≃)(a::AbstractArray{<:Tuple}, b::AbstractArray{<:Tuple})
+    size(a) == size(b) || return false
+    all(xy -> xy[1] ≃ xy[2], zip(a,b))
+end
+
+function isapproxfn(fn, args...; kwargs...)
+    isapprox(args...; kwargs...)
+end
+# Test against FiniteDifferences
+function test_scalar(f, x; rtol=1e-9, atol=1e-9, fdm=central_fdm(5, 1), kwargs...)
+    ∂x, = autodiff(ReverseHolomorphic, f, Active, Active(x))[1]
+
+    finite_diff = if typeof(x) <: Complex
+      RT = typeof(x).parameters[1]
+      (fdm(dx -> f(x+dx), RT(0)) - im * fdm(dy -> f(x+im*dy), RT(0)))/2
+    else
+      fdm(f, x)
+    end
+
+    @test isapproxfn((Enzyme.Reverse, f), ∂x, finite_diff; rtol=rtol, atol=atol, kwargs...)
+
+    if typeof(x) <: Integer
+        x = Float64(x)
+    end
+
+    if typeof(x) <: Complex
+        ∂re, = autodiff(Forward, f, Duplicated(x, one(typeof(x))))
+        ∂im, = autodiff(Forward, f, Duplicated(x, im*one(typeof(x))))
+        ∂x = (∂re - im*∂im)/2
+    else
+        ∂x, = autodiff(Forward, f, Duplicated(x, one(typeof(x))))
+    end
+
+    @test isapproxfn((Enzyme.Reverse, f), ∂x, finite_diff; rtol=rtol, atol=atol, kwargs...)
+
+end
+
+function test_matrix_to_number(f, x; rtol=1e-9, atol=1e-9, fdm=central_fdm(5, 1), kwargs...)
+    dx_fd = map(eachindex(x)) do i
+        fdm(x[i]) do xi
+            x2 = copy(x)
+            x2[i] = xi
+            f(x2)
+        end
+    end
+
+    dx = zero(x)
+    autodiff(Reverse, f, Active, Duplicated(x, dx))
+    @test isapproxfn((Enzyme.Reverse, f), reshape(dx, length(dx)), dx_fd; rtol=rtol, atol=atol, kwargs...)
+
+    dx_fwd = map(eachindex(x)) do i
+        dx = zero(x)
+        dx[i] = 1
+        ∂x = autodiff(Forward, f, Duplicated(x, dx))
+        isempty(∂x) ? zero(eltype(dx)) : ∂x[1]
+    end
+    @test isapproxfn((Enzyme.Forward, f), dx_fwd, dx_fd; rtol=rtol, atol=atol, kwargs...)
+end
+
